Some renames that in hindsight feel more organic.
